+++ src/components/project/review/review-dialog.tsx
@@ -1,8 +1,8 @@
 'use client';
 
 import { ITag, REVIEW_TAGS } from '@/constants/project';
-import { getProjectReview } from '@/lib/api/project/project-review';
 import { cn } from '@/lib/utils';
+import { getProjectReview } from '@/service/project/review';
 import { useEffect, useState } from 'react';
 import ConceptTag from '../../common/concept-tag';
 import Icon from '../../common/icon';
